Move Container to its final spot.
